FIX ugh only delete tmpfile if it exists

+++ attelo/cmd/decode.py
@@ -138,7 +138,8 @@ def delayed_main_for_harness(args, decoder, dpack, models):
     tmpfiles = [tmp_output_filename(args.output, d)
                 for d in groupings]
     for tmpfile in tmpfiles:
-        os.remove(tmpfile)
+        if fp.exists(tmpfile):
+            os.remove(tmpfile)
     for onedoc, indices in groupings.items():
         onepack = dpack.selected(indices)
         output = tmp_output_filename(args.output, onedoc)
